Manual merge of typecheckcleanup200810 branch as of rev 731346 in that branch, resolved all conflicts; with this BigDecimal is not the native type for many entity fields and is used for many calculations; some work is still needed in certain areas of the project but this covers the most critical ones; thanks especially to Scott for his work on this
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@731851 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/accounting/webapp/accounting/WEB-INF/actions/transaction/CaptureTransaction.groovy
@@ -54,7 +54,7 @@ if (orderPaymentPreference) {
 
     if (gatewayResponses) {
         latestAuth = gatewayResponses[0];
-        context.captureAmount = latestAuth.getDouble("amount");
+        context.captureAmount = latestAuth.getBigDecimal("amount");
     } else {
         // todo: some kind of error telling user to re-authorize
     }
